+++ tests/contrib/bots/test_choice_llm_bot.py
@@ -1,15 +1,15 @@
 from tokenize import Single
 from unittest import TestCase
 
-from dandy.handler.handler import Handler
+from dandy.processor.processor import BaseProcessor
 
 
 class TestChoiceLlmBot(TestCase):
     def test_single_choice_llm_bot_import(self):
         from dandy.contrib.bots import SingleChoiceLlmBot
-        self.assertTrue(type(SingleChoiceLlmBot) is type(Handler))
+        self.assertTrue(type(SingleChoiceLlmBot) is type(BaseProcessor))
 
     def test_multiple_choice_llm_bot_import(self):
         from dandy.contrib.bots import MultipleChoiceLlmBot
-        self.assertTrue(type(MultipleChoiceLlmBot) is type(Handler))
+        self.assertTrue(type(MultipleChoiceLlmBot) is type(BaseProcessor))