+++ tests/handler/test_handler.py
@@ -3,5 +3,5 @@ from unittest import TestCase
 
 class TestHandler(TestCase):
     def test_handler_import(self):
-        from dandy.handler.handler import Handler
-        self.assertTrue(type(Handler) is type(Handler))
+        from dandy.processor.processor import BaseProcessor
+        self.assertTrue(type(BaseProcessor) is type(BaseProcessor))